tut15 done

+++ client/src/components/WorkoutDetails.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 import '../index.css'
 
 // date-fns
@@ -7,25 +8,32 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 function WorkoutDetails({ workout }) {
   const { dispatch } = useWorkoutsContext()
+  const { user } = useAuthContext()
 
   const handleClick = async () => {
+    if (!user) {
+      return
+    }
     const response = await fetch('http://localhost:4000/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     const json = await response.json()
 
     if (response.ok) {
-      dispatch({type: 'DELETE_WORKOUT', payload: json})
+      dispatch({ type: 'DELETE_WORKOUT', payload: json })
     }
   }
 
   return (
     <div className='workout-details'>
-        <h4>{workout.title}</h4>
-        <p><strong>Load (kg): </strong>{workout.load}</p>
-        <p><strong>Reps: </strong>{workout.reps}</p>
-        <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
-        <span className='material-symbols-outlined' onClick={handleClick}>delete</span>
+      <h4>{workout.title}</h4>
+      <p><strong>Load (kg): </strong>{workout.load}</p>
+      <p><strong>Reps: </strong>{workout.reps}</p>
+      <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
+      <span className='material-symbols-outlined' onClick={handleClick}>delete</span>
     </div>
   )
 }
